+++ src/generator/files/src/__name@dasherize__/index.ts
@@ -9,24 +9,25 @@ import {strings} from '@angular-devkit/core';
 import {
   Rule,
   //Tree,
+  //SchematicContext
   apply,
   mergeWith,
   template,
   url,
   //chain,
 } from '@angular-devkit/schematics';
-//import blank from '@schematics/schematics/blank/factory';
 import {dasherize, classify, camelize} from '@angular-devkit/core/src/utils/strings';
-import {Schema as Options} from './schema';
+//import {Schema as Options} from './schema';
 
 
 const stringUtils = {dasherize, classify, camelize};
 
 // You don't have to export the function as default. You can also have more than one rule factory
 // per file.
-export function generator(options: Options): Rule {
-  /*return (tree: Tree) => {
-    return chain([blank({name: options.name})]);
+export function <%= camelize(name) %>(_options: any): Rule {
+  /*return (tree: Tree, _context: SchematicContext): Rule => {
+    return tree;
+    // return chain[example({name: options.name})]
   };*/
   return mergeWith(apply(url('./files'), [
     template({