+++ src/generator/files/src/__name@dasherize__/index__dot__spec.ts
@@ -1,0 +1,32 @@
+/**
+ * @license
+ * Tomasz Gorka <https://tomasz.gorka.org.pl>
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/tgorka/cliberry-schematics/LICENSE
+ */
+//import { Tree } from '@angular-devkit/schematics';
+import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
+import * as path from 'path';
+import {Schema as Options} from './schema';
+
+
+const collectionPath = path.join(__dirname, '../collection.json');
+const schematicName = '<%= dasherize(name) %>';
+
+describe(schematicName, () => {
+  const schematicRunner = new SchematicTestRunner('schematics', collectionPath);
+  const defaultOptions: Options = {
+    name: 'foo',
+    description: 'foo description',
+  };
+
+  it('should create all files of a project', () => {
+    const options = {...defaultOptions};
+
+    const tree = schematicRunner.runSchematic(schematicName, options);
+    const files = tree.files;
+    expect(files.indexOf('/src/<%= dasherize(name) %>/.gitkeep')).toBeGreaterThanOrEqual(0);
+  });
+
+});