+++ src/generator/files/src/__name@dasherize__/schema.d.ts
@@ -1,0 +1,17 @@
+/**
+ * @license
+ * Tomasz Gorka <https://tomasz.gorka.org.pl>
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/tgorka/cliberry-schematics/LICENSE
+ */
+export interface Schema {
+  /**
+   * The name of the generator.
+   */
+  name: string;
+  /**
+   * Short description of the generator.
+   */
+  description: string;
+}